# packages/contracts/openapi/trillo.v1.yaml
openapi: 3.0.3
info:
  title: Trillo Tasks API
  version: 1.0.0
  description: API contract for the Task Manager kanban application.
servers:
  - url: /api/v1
paths:
  /projects:
    get:
      summary: List projects
      responses:
        '200':
          description: Projects returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProjectsResponse'
    post:
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Project name conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /projects/{projectId}:
    delete:
      summary: Delete project and its board data
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 64
      responses:
        '204':
          description: Project deleted
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /tasks:
    get:
      summary: List tasks by board
      parameters:
        - in: query
          name: boardId
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 64
      responses:
        '200':
          description: Tasks returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTasksResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /tasks/{taskId}/status:
    patch:
      summary: Move task to another status column
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveTaskStatusRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /tasks/{taskId}:
    patch:
      summary: Update task details
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Conflict with linked tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      summary: Delete a task
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task deleted
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Conflict with linked tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  schemas:
    Project:
      type: object
      required:
        - id
        - name
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          minLength: 2
          maxLength: 64
        name:
          type: string
          minLength: 2
          maxLength: 120
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskStatus:
      type: string
      enum: [todo, in_progress, done]
    TaskPriority:
      type: string
      enum: [low, medium, high]
    TaskType:
      type: string
      enum: [task, epic]
    Task:
      type: object
      required:
        - id
        - boardId
        - title
        - description
        - category
        - priority
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        boardId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        priority:
          $ref: '#/components/schemas/TaskPriority'
        status:
          $ref: '#/components/schemas/TaskStatus'
        taskType:
          $ref: '#/components/schemas/TaskType'
        epicId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ListTasksResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        meta:
          type: object
          required: [boardId, total]
          properties:
            boardId:
              type: string
            total:
              type: integer
              minimum: 0
    ListProjectsResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        meta:
          type: object
          required: [total]
          properties:
            total:
              type: integer
              minimum: 0
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
    CreateTaskRequest:
      type: object
      required: [boardId, title, category]
      properties:
        boardId:
          type: string
          minLength: 2
          maxLength: 64
        title:
          type: string
          minLength: 3
          maxLength: 140
        description:
          type: string
          minLength: 0
          maxLength: 4000
        category:
          type: string
          minLength: 2
          maxLength: 32
        priority:
          $ref: '#/components/schemas/TaskPriority'
        taskType:
          $ref: '#/components/schemas/TaskType'
        epicId:
          type: string
          format: uuid
          nullable: true
    MoveTaskStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/TaskStatus'
    UpdateTaskRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 140
        description:
          type: string
          nullable: true
          minLength: 0
          maxLength: 4000
        category:
          type: string
          minLength: 2
          maxLength: 32
        priority:
          $ref: '#/components/schemas/TaskPriority'
        taskType:
          $ref: '#/components/schemas/TaskType'
        epicId:
          type: string
          format: uuid
          nullable: true
    TaskResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Task'
    ProjectResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Project'
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties:
                type: string